OneDrive/Bureau/userModule-main/userModule-main/Backend/face-service/requirements.txt
starlette==0.47.3
uvicorn==0.35.0
python-multipart==0.0.20
Pillow==11.3.0
deepface==0.0.93
